Move isChildNode to lib/dom
